Share OpenApi testing functionality between beacon api and validator api.
There's starting to be a reasonable set of resources around checking openapi specifications, so it's worth consolidating the methodology and sharing the code.
