+++ data/beaconrestapi/build.gradle
@@ -30,14 +30,15 @@ dependencies {
 
     testImplementation testFixtures(project(':bls'))
     testImplementation testFixtures(project(':storage'))
-testImplementation testFixtures(project(':ethereum:spec'))
+    testImplementation testFixtures(project(':ethereum:spec'))
     testImplementation project(':ethereum:networks')
     testImplementation testFixtures(project(':infrastructure:async'))
+    testImplementation testFixtures(project(':infrastructure:restapi'))
 
     testCompileOnly 'io.libp2p:jvm-libp2p-minimal'
 
     integrationTestImplementation testFixtures(project(':bls'))
-integrationTestImplementation testFixtures(project(':ethereum:spec'))
+    integrationTestImplementation testFixtures(project(':ethereum:spec'))
     integrationTestImplementation project(':ethereum:weaksubjectivity')
     integrationTestImplementation testFixtures(project(':ethereum:core'))
