--- data/beaconrestapi/src/integration-test/resources/tech/pegasys/teku/beaconrestapi/beacon/paths_listing.txt
@@ -1,55 +1,0 @@
-_eth_v1_beacon_blocks.json
-_eth_v1_beacon_blocks_{block_id}.json
-_eth_v1_beacon_blocks_{block_id}_attestations.json
-_eth_v1_beacon_blocks_{block_id}_root.json
-_eth_v1_beacon_genesis.json
-_eth_v1_beacon_headers.json
-_eth_v1_beacon_headers_{block_id}.json
-_eth_v1_beacon_pool_attestations.json
-_eth_v1_beacon_pool_attester_slashings.json
-_eth_v1_beacon_pool_proposer_slashings.json
-_eth_v1_beacon_pool_sync_committees.json
-_eth_v1_beacon_pool_voluntary_exits.json
-_eth_v1_beacon_states_{state_id}_committees.json
-_eth_v1_beacon_states_{state_id}_finality_checkpoints.json
-_eth_v1_beacon_states_{state_id}_fork.json
-_eth_v1_beacon_states_{state_id}_root.json
-_eth_v1_beacon_states_{state_id}_sync_committees.json
-_eth_v1_beacon_states_{state_id}_validator_balances.json
-_eth_v1_beacon_states_{state_id}_validators.json
-_eth_v1_beacon_states_{state_id}_validators_{validator_id}.json
-_eth_v1_config_deposit_contract.json
-_eth_v1_config_fork_schedule.json
-_eth_v1_config_spec.json
-_eth_v1_debug_beacon_heads.json
-_eth_v1_debug_beacon_states_{state_id}.json
-_eth_v1_events.json
-_eth_v1_node_health.json
-_eth_v1_node_identity.json
-_eth_v1_node_peer_count.json
-_eth_v1_node_peers.json
-_eth_v1_node_peers_{peer_id}.json
-_eth_v1_node_syncing.json
-_eth_v1_node_version.json
-_eth_v1_validator_aggregate_and_proofs.json
-_eth_v1_validator_aggregate_attestation.json
-_eth_v1_validator_attestation_data.json
-_eth_v1_validator_beacon_committee_subscriptions.json
-_eth_v1_validator_blocks_{slot}.json
-_eth_v1_validator_contribution_and_proofs.json
-_eth_v1_validator_duties_attester_{epoch}.json
-_eth_v1_validator_duties_proposer_{epoch}.json
-_eth_v1_validator_duties_sync_{epoch}.json
-_eth_v1_validator_liveness.json
-_eth_v1_validator_sync_committee_contribution.json
-_eth_v1_validator_sync_committee_subscriptions.json
-_eth_v2_beacon_blocks_{block_id}.json
-_eth_v2_debug_beacon_states_{state_id}.json
-_eth_v2_validator_blocks_{slot}.json
-_teku_v1_admin_liveness.json
-_teku_v1_admin_log_level.json
-_teku_v1_admin_readiness.json
-_teku_v1_beacon_blocks_{block_id}_state.json
-_teku_v1_beacon_blocks_{slot}.json
-_teku_v1_beacon_states_{state_id}.json
-_teku_v1_nodes_peer_scores.json
