--- data/beaconrestapi/src/integration-test/resources/tech/pegasys/teku/beaconrestapi/beacon/schema_listing.txt
@@ -1,117 +1,0 @@
-AggregateAndProof.json
-Attestation.json
-AttestationData.json
-AttesterDuty.json
-AttesterSlashing.json
-BeaconBlockAltair.json
-BeaconBlock.json
-BeaconBlockBody.json
-BeaconBlockBodyAltair.json
-BeaconBlockBodyMerge.json
-BeaconBlockHeader.json
-BeaconBlockMerge.json
-BeaconBlockPhase0.json
-BeaconCommitteeSubscriptionRequest.json
-BeaconStateAltair.json
-BeaconStateMerge.json
-BeaconStatePhase0.json
-BlockHeader.json
-Bytes.json
-Bytes32.json
-ChainHead.json
-Checkpoint.json
-ContributionAndProof.json
-Deposit.json
-DepositContract.json
-DepositData.json
-EpochCommitteeResponse.json
-Eth1Data.json
-ExecutionPayload.json
-ExecutionPayloadHeader.json
-FinalityCheckpointsResponse.json
-Fork.json
-GenesisData.json
-GetAggregatedAttestationResponse.json
-GetAllBlocksAtSlotResponse.json
-GetAttestationDataResponse.json
-GetAttestationsResponse.json
-GetAttesterSlashingsResponse.json
-GetBlockAttestationsResponse.json
-GetBlockHeaderResponse.json
-GetBlockHeadersResponse.json
-GetBlockResponse.json
-GetBlockResponseV2.json
-GetBlockRootResponse.json
-GetChainHeadsResponse.json
-GetDepositContractResponse.json
-GetForkScheduleResponse.json
-GetGenesisResponse.json
-GetNewBlockResponse.json
-GetNewBlockResponseV2.json
-GetPeerCountResponse.json
-GetPeerScoresResponse.json
-GetProposerDutiesResponse.json
-GetProposerSlashingsResponse.json
-GetSpecResponse.json
-GetStateCommitteesResponse.json
-GetStateFinalityCheckpointsResponse.json
-GetStateForkResponse.json
-GetStateResponse.json
-GetStateResponseV2.json
-GetStateRootResponse.json
-GetStateSyncCommitteesResponse.json
-GetStateValidatorBalancesResponse.json
-GetStateValidatorResponse.json
-GetStateValidatorsResponse.json
-GetSyncCommitteeContributionResponse.json
-GetVoluntaryExitsResponse.json
-Identity.json
-IdentityResponse.json
-IndexedAttestation.json
-LogLevel.json
-Metadata.json
-Peer.json
-PeerCount.json
-PeerResponse.json
-PeerScore.json
-PeersResponse.json
-PendingAttestation.json
-PostAttesterDutiesResponse.json
-PostDataFailure.json
-PostDataFailureResponse.json
-PostSyncDutiesResponse.json
-PostValidatorLivenessResponse.json
-ProposerDuty.json
-ProposerSlashing.json
-Root.json
-SignedAggregateAndProof.json
-SignedBeaconBlock.json
-SignedBeaconBlockAltair.json
-SignedBeaconBlockHeader.json
-SignedBeaconBlockMerge.json
-SignedBeaconBlockPhase0.json
-SignedBlock.json
-SignedContributionAndProof.json
-SignedVoluntaryExit.json
-State.json
-StateSyncCommittees.json
-SyncAggregate.json
-SyncCommittee.json
-SyncCommitteeContribution.json
-SyncCommitteeDuty.json
-SyncCommitteeMessage.json
-SyncCommitteeSubnetSubscription.json
-Syncing.json
-SyncingResponse.json
-UInt256.json
-UInt64.json
-Unit.json
-UnsignedBlock.json
-Validator.json
-ValidatorBalanceResponse.json
-ValidatorLivenessAtEpoch.json
-ValidatorLivenessRequest.json
-ValidatorResponse.json
-Version.json
-VersionResponse.json
-VoluntaryExit.json
