--- validator/client/src/test/resources/tech/pegasys/teku/validator/client/restapi/apis/PostKeys.json
@@ -1,58 +1,0 @@
-{
-  "post": {
-    "operationId": "ImportKeystores",
-    "summary": "Import Keystores",
-    "description": "Import keystores generated by the Eth2.0 deposit CLI tooling.",
-    "requestBody": {
-      "content": {
-        "application/json": {
-          "schema": {
-            "$ref": "#/components/schemas/PostKeysRequest"
-          }
-        }
-      }
-    },
-    "responses": {
-      "200": {
-        "description": "Success response",
-        "content": {
-          "application/json": {
-            "schema": {
-              "$ref": "#/components/schemas/PostKeysResponse"
-            }
-          }
-        }
-      },
-      "401": {
-        "description": "Unauthorized, no token is found",
-        "content": {
-          "application/json": {
-            "schema": {
-              "$ref": "#/components/schemas/HttpErrorResponse"
-            }
-          }
-        }
-      },
-      "403": {
-        "description": "Forbidden, a token is found but is invalid",
-        "content": {
-          "application/json": {
-            "schema": {
-              "$ref": "#/components/schemas/HttpErrorResponse"
-            }
-          }
-        }
-      },
-      "500": {
-        "description": "Internal server error",
-        "content": {
-          "application/json": {
-            "schema": {
-              "$ref": "#/components/schemas/HttpErrorResponse"
-            }
-          }
-        }
-      }
-    }
-  }
-}
