+++ validator/client/src/test/resources/tech/pegasys/teku/validator/client/restapi/schema/HttpErrorResponse.json
@@ -1,0 +1,12 @@
+{
+  "title" : "HttpErrorResponse",
+  "type" : "object",
+  "properties" : {
+    "status" : {
+      "type" : "number"
+    },
+    "message" : {
+      "type" : "string"
+    }
+  }
+}
